requirements.txt
psutil
discord.py
logging